requirements.txt
openai
requests
feedparser
python-dotenv